+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/WalletMain.kt
@@ -58,6 +58,7 @@ class WalletMain(
     val settingsRepository: SettingsRepository,
     val sessionService: SessionService,
     val capabilitiesService: CapabilitiesService,
+    val credentialValidityService: CredentialValidityService,
 ) {
     val appReady = MutableStateFlow<Boolean?>(null)
 
@@ -74,6 +75,7 @@ class WalletMain(
 
     init {
         startListeningForNewCredentialsDCAPI()
+        credentialValidityService.startChecking()
         if (keyMaterial.keyMaterial is FallBackKeyMaterial) {
             Napier.e("FallBackKeyMaterial: ${keyMaterial.keyMaterial.reason}")
         }