+++ shared/src/commonMain/kotlin/ui/navigation/routes/WalletRoutes.kt
@@ -214,3 +214,6 @@ data class CapabilitiesRoute(val prerequisitesSerialized: String) : Route() {
     val prerequisites: Set<RoutePrerequisites>
         get() = vckJsonSerializer.decodeFromString(prerequisitesSerialized)
 }
+
+@Serializable
+object RefreshCenterRoute